requirements.txt
absl-py==0.9.0
tensorflow==2.3.0
pose_format==0.0.1
numpy>=1.13.0
